requirements.txt
openai>=1.12.0
anthropic>=0.18.1
python-dotenv>=1.0.1
pydantic>=2.6.1
aiohttp>=3.9.3
python-multipart>=0.0.9  # For handling multipart form data (images)
typing-extensions>=4.9.0
colorama>=0.4.6
streamlit>=1.32.0
asyncio>=3.4.3
aiofiles>=23.2.1
python-json-logger>=2.0.7
